fix: move localStorage-dependent changelog fns from @aggy/lib to @aggy/ui
packages/lib is pure domain logic with lib:["ESNext"] — no DOM types.
hasUnreadChanges/markChangelogSeen use localStorage (browser API), so
they belong in packages/ui which already has lib:["ESNext","DOM",...].

- packages/lib/src/changelog.ts: keep only ChangelogEntry type + data
- packages/lib/tests/changelog.test.ts: removed (no functions left to test)
- packages/ui/src/lib/changelog.ts: new — hasUnreadChanges/markChangelogSeen
- packages/ui/tests/changelog.test.ts: new — tests moved here with polyfill
- ChangelogButton.svelte: import browser fns from ../../lib/changelog
