+++ packages/lib/src/changelog.ts
@@ -8,15 +8,3 @@ export interface ChangelogEntry {
 }
 
 export const changelog: ChangelogEntry[] = entries;
-
-const STORAGE_KEY = "aggy-changelog-seen";
-
-export function hasUnreadChanges(): boolean {
-  if (changelog.length === 0) return false;
-  return localStorage.getItem(STORAGE_KEY) !== changelog[0].version;
-}
-
-export function markChangelogSeen(): void {
-  if (changelog.length === 0) return;
-  localStorage.setItem(STORAGE_KEY, changelog[0].version);
-}
