+++ packages/ui/src/lib/changelog.ts
@@ -1,0 +1,13 @@
+import { changelog } from "@aggy/lib/changelog";
+
+const STORAGE_KEY = "aggy-changelog-seen";
+
+export function hasUnreadChanges(): boolean {
+  if (changelog.length === 0) return false;
+  return localStorage.getItem(STORAGE_KEY) !== changelog[0].version;
+}
+
+export function markChangelogSeen(): void {
+  if (changelog.length === 0) return;
+  localStorage.setItem(STORAGE_KEY, changelog[0].version);
+}
